Basic website to search a plant database using HTML and PHP
This website was developed for the Barton plants database.
The front end was developed using HTML and CSS. The CSS template was
borrowed from templated.co. The underlying database was managed in
MySQL. The program uses PHP scripts to communicate with the database
server.

The website provides users the feature to search specific plants from
the database using keywords. Users can also search by color of the plant
or month in which they bloom.

This website was developed as a part of a class project for database
management in Spring 2015.
Link:

http://corvette.ischool.utexas.edu/htdocs/bartonplants/bartonplants/index.php